main: split entry.c into multiple files to allow output handler plugable
New interface:

    typedef int (* writeEntryFunc) (MIO * mio, const tagEntryInfo *const tag, void *data);
    typedef void * (* preWriteEntryFunc) (MIO * mio);
    typedef void (* postWriteEntryFunc)  (MIO * mio, const char* filename, void *data);

    extern void setTagWriter (writeEntryFunc func,
			      preWriteEntryFunc preFunc,
			      postWriteEntryFunc postFunc);

Example:

    extern int writeEtagsEntry (MIO * mio, const tagEntryInfo *const tag, void *data);
    extern void *beginEtagsFile (MIO * mio);
    extern void  endEtagsFile   (MIO * mio, const char* filename, void *data);
