+++ win32/ctags_vs2013.vcxproj
@@ -146,6 +146,9 @@
     <ClCompile Include="..\parsers\objc.c" />
     <ClCompile Include="..\parsers\ocaml.c" />
     <ClCompile Include="..\main\options.c" />
+    <ClCompile Include="..\main\output-ctags.c" />
+    <ClCompile Include="..\main\output-etags.c" />
+    <ClCompile Include="..\main\output-xref.c" />
     <ClCompile Include="..\main\parse.c" />
     <ClCompile Include="..\parsers\pascal.c" />
     <ClCompile Include="..\parsers\perl.c" />
@@ -223,6 +226,7 @@
     <ClInclude Include="..\main\main.h" />
     <ClInclude Include="..\main\nestlevel.h" />
     <ClInclude Include="..\main\options.h" />
+    <ClInclude Include="..\main\output.h" />
     <ClInclude Include="..\main\parse.h" />
     <ClInclude Include="..\main\parsers.h" />
     <ClInclude Include="..\main\ptag.h" />
