+++ win32/ctags_vs2013.vcxproj.filters
@@ -241,6 +241,15 @@
     <ClCompile Include="..\main\options.c">
       <Filter>Source Files\Main</Filter>
     </ClCompile>
+    <ClCompile Include="..\main\output-ctags.c">
+      <Filter>Source Files\Main</Filter>
+    </ClCompile>
+    <ClCompile Include="..\main\output-etags.c">
+      <Filter>Source Files\Main</Filter>
+    </ClCompile>
+    <ClCompile Include="..\main\output-xref.c">
+      <Filter>Source Files\Main</Filter>
+    </ClCompile>
     <ClCompile Include="..\main\parse.c">
       <Filter>Source Files\Main</Filter>
     </ClCompile>
@@ -380,6 +389,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\main\options.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="..\main\output.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\main\parse.h">
